ENH: Split and move events.py module into events directory
Split the weekly reminders and the slash commands into separate modules. slash_commands.py and weekly_reminders.py.
